# dtos/user_dto.go
package dtos

import (
	"github.com/atqamz/kogase-backend/models"
)

type CreateUserRequest struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=6"`
	Name     string `json:"name" binding:"required"`
}

type CreateUserResponse struct {
	UserID string `json:"user_id"`
	Email  string `json:"email"`
	Name   string `json:"name"`
}

type GetUsersResponse struct {
	Users []GetUserResponse `json:"users"`
}

type GetUserResponse struct {
	UserID string `json:"user_id"`
	Email  string `json:"email"`
	Name   string `json:"name"`
}

type GetUserResponseDetail struct {
	GetUserResponse
	Projects []models.Project `json:"projects"`
}

type UpdateUserRequest struct {
	Email    string `json:"email" binding:"omitempty,email"`
	Name     string `json:"name" binding:"omitempty"`
	Password string `json:"password" binding:"omitempty"`
}

type UpdateUserResponse struct {
	Email string `json:"email"`
	Name  string `json:"name"`
}

type DeleteUserResponse struct {
	Message string `json:"message"`
}